Fix libnrnmech lib name with neuron+cmake+binary

+++ var/spack/repos/builtin/packages/sim-model/package.py
@@ -151,7 +151,7 @@ class SimModel(Package):
         # Install special
         shutil.copy(join_path(arch, 'special'), prefix.bin)
 
-        if self.spec.satisfies('^neuron~binary'):
+        if self.spec.satisfies('^neuron~binary') or self.spec.satisfies('^neuron+binary+cmake'):
             lib_suffix = ".so"
             # Install libnrnmech - might have several links
             if self.spec.satisfies('^neuron+cmake'):
@@ -165,18 +165,22 @@ class SimModel(Package):
                           recursive=False):
                 if not os.path.islink(f):
                     bname = os.path.basename(f)
-                    lib_dst = prefix.lib.join(
-                        bname[:bname.find('.')] + self.lib_suffix + '.so')
+                    if self.spec.satisfies('^neuron+binary+cmake'):
+                        lib_dst = prefix.lib.join(bname)
+                    else:
+                        lib_dst = prefix.lib.join(
+                            bname[:bname.find('.')] + self.lib_suffix + '.so')
                     shutil.move(f, lib_dst)  # Move so its not copied twice
                     break
             else:
                 raise Exception('No libnrnmech found')
 
-            # Patch special for the new libname
-            which('sed')('-i.bak',
-                         's#-dll .*#-dll %s "$@"#' % lib_dst,
-                         prefix.bin.special)
-            os.remove(prefix.bin.join('special.bak'))
+            if self.spec.satisfies('^neuron~binary'):
+                # Patch special for the new libname
+                which('sed')('-i.bak',
+                             's#-dll .*#-dll %s "$@"#' % lib_dst,
+                             prefix.bin.special)
+                os.remove(prefix.bin.join('special.bak'))
 
     def _install_src(self, prefix):
         """Copy original and translated c mods
